Fix changeStatusMessage-test flakyness
Depending on simulator, when scrolling down the form, it will touch the
Switch and then the test fails, because the Switch causes the app to
patch wrong data.

Added a option to set coordinates (x, y) on scroll position, and
keeping (0.1, 0.2) as default

Tap the save-button again because of multiline input-field needs to be
focused out of. Scrolling is not enough.